fix(web): top bar tweaks
